fix the issues with different classloaders
The problem apparently occurs, because the ecore2owl.core was in the libs.jar and thus had a totally different classloader. Therefore, there were multiple problems like no registered ResourceFactories or being unable to call a function with a parameter class loaded by a different classloader. Currently, the solution is to minimize the stuff that needs to be loaded from the libs.jar. As a consequence, I transformed ecore2owl.core and OntologyAccess into plugins and introduced the dependencies-collector (maven-only) and dependencies-provider (loading the libs.jar and providing the stuff in there).